Updating python scripts for AVIF generation The optimization process now includes AVIF generation, using `cavif-rs` Also, upon finishing the filename is saved to prevent re-optimizing

+++ lib/toolset/optimize_images.py
@@ -1,12 +1,22 @@
 #!/usr/bin/python
 
-from pathlib import Path
+import json
+import subprocess
 import sys
+
+from pathlib import Path
 
 from PIL import Image
 
 
-def optimize_images(path: str, quality: int=85):
+def optimize_images(path: str):
+    # Load list of already optimized images (if it exists)
+    try:
+        with open('storage/optimized_images.json', 'r') as file:
+            data = json.load(file)
+    except EnvironmentError:
+        data = []
+
     # Extensions to look for
     extensions = [
         '.png',
@@ -16,30 +26,60 @@ def optimize_images(path: str, quality: int=85):
 
     # Get this party started
     for file in Path(path).rglob('*'):
-        if file.suffix.lower() in extensions:
+        if file.name not in data and file.suffix.lower() in extensions:
             try:
-                # Create optimized JPEG image
                 image = Image.open(file)
                 file.unlink()
+
+                # Create optimized version of original image
+                print('Processing {} ..'.format(file.name))
 
                 if image.mode != 'RGB':
                     image = image.convert('RGB')
 
-                jpg_file = file.with_suffix('.jpg')
+                if file.suffix == '.png':
+                    # If original image format is PNG:
+                    # Make it as small as possible
+                    image.save(file, format='PNG', optimize=True)
+                else:
+                    # If original image format is JPEG:
+                    # (1) Make it progressive
+                    # (2) at maximum quality
+                    # (3) but strip metadata
+                    # (4) and set chroma subsampling mode to `420`
+                    image.save(file, format='JPEG', quality='keep', progressive=True, optimize=True, subsampling=2)
 
-                print('Processing {} ..'.format(file.name))
-                image.save(jpg_file, format='JPEG', quality=quality, progressive=True, optimize=True, sampling=2)
+                # Close file pointer
+                image.close()
 
                 # Create optimized WebP image
-                image = Image.open(jpg_file)
-                webp_file = jpg_file.with_suffix('.webp')
+                image = Image.open(file)
+                webp_file = file.with_suffix('.webp')
 
+                # Maximum compression (better, but slower)
                 print('Creating {} ..'.format(webp_file.name))
                 image.save(webp_file, format='WebP', method=6)
 
+                # Close file pointer
+                image.close()
+
+                # Create optimized AVIF image
+                # See https://github.com/kornelski/cavif-rs
+                avif_file = file.with_suffix('.avif')
+
+                # Overwrite if it already exists (must be enabled)
+                print('Creating {} ..'.format(avif_file.name))
+                subprocess.call(['cavif', str(file), '--overwrite', '--quiet'])
+
+                # Add image to list of optimized images ..
+                data.append(file.name)
+
+                # .. and save the updated list
+                with open('storage/optimized_images.json', 'w') as file:
+                    json.dump(data, file)
+
             except Exception as error:
                 print(error)
-                continue
 
 
 if __name__ == '__main__':
